initializing the graphic editor
we have different objects (line, circle, rectangle, triangle, freeform, polygon) that i want to draw in the canvas. The objects should be graphical interactive so that you can define them with the mouse. At first I only had to implement drawing a line, a circle, a rectangle and the freehand line. In the GUI there was a color botton which you would use to change the color before drawing the new object. Then there is another checkbox "Show only outline" which clicked only shows the outline of objects. Another function is also the "Clear Canvas" button which clears the whole canvas when clicked